PyExec(): moved to .common; implemented valid_symbols
